@@ -2,5 +2,8 @@ const admin = require('firebase-admin');
 
 admin.initializeApp();
 
-exports.user = require('./api/user.api');
-exports.userTriggers = require('./triggers/user.trigger');
+// APIs
+exports.carryOver = require('./api/helloworld.api');
+
+// Triggers
+// exports.entryTriggers = require('./triggers/helloworld.trigger');
