@@ -1,51 +1,44 @@
-import './Spinner.css';
+import Backdrop from '@mui/material/Backdrop';
+import CircularProgress from '@mui/material/CircularProgress';
 import { useEffect, useState } from 'react';
-import { DotLoader } from 'react-spinners';
-import { css } from '@emotion/react';
+import ErrorService from '../services/ErrorService';
+import { useAppDispatch } from '../store/Hooks';
 
 interface ContainerProps {
   show: boolean,
 }
 
-const override = css`
-  display: block;
-  margin: 0 auto;
-`;
-
 const Spinner: React.FC<ContainerProps> = ({
   show,
 }) => {
-  const [showSpinner, setShowSpinner] = useState(false);
-  const [activeClass, setActiveClass] = useState('');
+  const dispatch = useAppDispatch();
 
+  const [open, setOpen] = useState(false);
+
+  // eslint-disable-next-line consistent-return
   useEffect(() => {
     try {
       if (show) {
-        // Prevent the spinner from stuttering if load time is not enough
-        const fadeInTimer = setTimeout(() => {
-          setShowSpinner(true);
-          setActiveClass('active');
-        }, 600);
-        return () => clearTimeout(fadeInTimer);
-      }
+        const timer1 = setTimeout(() => {
+          setOpen(show);
+        }, 500);
 
-      setActiveClass('');
-      const fadeOutTimer = setTimeout(() => {
-        setShowSpinner(false);
-      }, 600);
-      return () => clearTimeout(fadeOutTimer);
+        return () => {
+          clearTimeout(timer1);
+        };
+      }
     } catch (e: any) {
-      setShowSpinner(false);
-      setActiveClass('');
-      console.error(e);
-      return e;
+      ErrorService.handleError(e, dispatch);
     }
   }, [show]);
 
   return (
-    <div className={`spinner__container ${activeClass}`}>
-      <DotLoader css={override} loading={showSpinner} size={150}/>
-    </div>
+    <Backdrop
+      sx={{ color: '#fff', zIndex: (theme) => theme.zIndex.drawer + 1 }}
+      open={open}
+    >
+      <CircularProgress color="inherit" />
+    </Backdrop>
   );
 };
 
