@@ -1,11 +1,11 @@
 import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit';
 import { userSlice } from './slices/UserSlice';
-import { spinnerSlice } from './slices/SpinnerSlice';
+import { alertSlice } from './slices/AlertSlice';
 
 const store = configureStore({
   reducer: {
     user: userSlice.reducer,
-    loader: spinnerSlice.reducer,
+    alert: alertSlice.reducer,
   },
   middleware: getDefaultMiddleware({
     serializableCheck: false,
