@@ -1,10 +1,10 @@
 /* eslint-disable no-param-reassign */
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import SupportedLanguages from '../../models/enums/SupportedLanguages';
 
 const initialState = {
   user: null,
-  firebaseUser: null,
-  platform: null,
+  language: SupportedLanguages.DEFAULT,
 };
 
 export const userSlice = createSlice({
@@ -14,25 +14,15 @@ export const userSlice = createSlice({
     updateUser(state, action: PayloadAction<any>) {
       state.user = action.payload;
     },
-    updateFirebaseUser(state, action: PayloadAction<any>) {
-      state.firebaseUser = action.payload;
-    },
-    updatePlatform(state, action: PayloadAction<any>) {
-      state.platform = action.payload;
-    },
-    logout(state) {
-      state.platform = null;
-      state.user = null;
-      state.firebaseUser = null;
+    updateLanguage(state, action: PayloadAction<any>) {
+      state.language = action.payload;
     },
   },
 });
 
 export const {
   updateUser,
-  updateFirebaseUser,
-  updatePlatform,
-  logout,
+  updateLanguage,
 } = userSlice.actions;
 
 export default userSlice;
